Add GameController factory and bug fixes
The GameController contains the logic to play the game and is playable
within the browser console. The function also returns two additional
functions, getCurrentPlayer and getGameStatus, which aren't needed to
play the game in the console but will be useful for the browser UI.

Bug Fixes:
- The playerInput() function inside the Gameboard factory didn't work.
The if conditions inside the function weren't looking up the value of
the selected cell, and so the code was never being executed. Fixed this
by correctly referencing the cell values within the if condition.
- Fixed the cell function to add a player token. The code was written
before building the player objects. As such, it was written to update
the cell values to be the player object rather than the players token.
Fixed this by extracting the token from the player object.

Additional Notes:
- Pseudocode has been left in the script for now to help explain the
thought process behind the code. When the script is finalised, the
pseudocode can be removed.