Fix game breaking token override bug
The playerInput function inside the Gameboard factory was initially set
up to end if the board cell was already populated with a player token.
This was intended to prevent a player from being able to override an
already populated cells value. However, when playRound was called, if a
player made an invalid selection, playRound would continue to run even
though the move was invalid and the player hadn't changed the game
state. The game would then switch to the next player, effectively making
a player miss a turn if they made an invlaid selection.
To fix this, I've introduced a new variable, invalidMove, inside the
Gameboard factory. This gets set to 1 if a player makes an invalid move.
The GameController can then call a function, getInvalidMove, which will
retrieve this variable. The variable is then used within the playRound
function to see whether a player made an invalid move. If the did, then
playRound ends early forcing the player to make another selection.