Fix win condition bug for tie games
winConditions was previously set up to declare a tie if the game board
was full. The bug was that it had been set up in a way that *always*
declated a tie when the board was full, even if a player one on the last
tile.

To fix this, and hopefully make the code clearer, I've removed the tie
game logic from the winConditions function and created a separate
tieCondition function. I've modified the logic inside this new function
to not only check if the board is full, but also to check the gameStatus
variable to see if a player has won. Now a tie game is only called if
the board if full and no player has won on the last move.